Compute distance between location

+++ app/src/main/java/studio/bz_soft/freightforwarder/data/models/DistanceModel.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class DistanceModel(
     @SerializedName("user_id") val userId :Int?,
     @SerializedName("work_shift") val workShift :String?,
-    @SerializedName("distance") val distance: Double?
+    @SerializedName("distance") val distance: Float?
 )
